organize a bit
