--- src/node/utils/slash.ts
@@ -1,3 +1,0 @@
-export function slash(p: string): string {
-  return p.replace(/\\/g, '/')
-}
